+++ CHANGELOG.md
@@ -5,11 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `FreedesktopPortalProxyResolver` to lookup proxies via flatpak portal.
+
+### Changed
+- Simplify `env` module after removal of `ProxyResolver`.
+
 ### Removed
 
 - Drop trait `ProxyResolver`.
   Introduced prematurely; instead wait until a common API matures.
-- Simplify `env` module after removal of `ProxyResolver`.
 - Make Gio proxy resolver asynchronous.
 
 ## [0.2.0] – 2023-03-02